third_party/WebKit/LayoutTests/editing/selection/mouse/click-user-select-all-textarea.html

<!doctype HTML>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<textarea id="textarea" style="-webkit-user-select:all"></textarea>
<script>
test(function(){
    assert_true(window.hasOwnProperty('eventSender'), 'this test requires window.eventSender');
    eventSender.mouseMoveTo(textarea.offsetLeft + 5, textarea.offsetTop + 5);
    eventSender.mouseDown();
    eventSender.mouseUp();
    eventSender.keyDown('f');
    eventSender.keyDown('o');
    eventSender.keyDown('o');
    assert_equals(textarea.value, 'foo');
    },
'User can ipnut in -webkit-user-select:all textarea');
</script>
